requirements.txt
asgiref==3.5.2
asttokens==2.2.1
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
cloudinary==1.30.0
colorama==0.4.6
comm==0.1.2
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
distlib==0.3.6
dj-database-url==1.0.0
dj3-cloudinary-storage==0.0.6
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
django-heroku==0.3.1
djangorestframework==3.14.0
entrypoints==0.4
executing==1.2.0
filelock==3.8.0
fonttools==4.38.0
gunicorn==20.1.0
idna==3.4
ipykernel==6.19.4
ipython==8.7.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==7.4.8
jupyter_core==5.1.1
kiwisolver==1.4.4
Markdown==3.4.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.0
packaging==22.0
pandas==1.5.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.2
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.5
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
pywin32==305
pyzmq==24.0.1
requests==2.28.1
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.8.0
tzdata==2022.4
unicorn==2.0.1
urllib3==1.26.12
virtualenv==20.16.6
wcwidth==0.2.5
whitenoise==6.2.0
